requirements.txt
selenium
fastapi
uvicorn
jinja2
python-multipart
python-dotenv
bs4
html2text
# alternat
gunicorn